[refactor/#44] 마이페이지 조회 시 이번 주 지난 주 성공률 정보 추가

+++ src/main/java/neordinary/dofarming/api/controller/user/dto/response/GetUserRes.java
@@ -16,7 +16,7 @@ public class GetUserRes {
     private String nickname;
     private String profileImageUrl;
     private List<GetPointRes> pointList;
-    private Integer lastWeekMissionSuccessRate;
-    private Integer thisWeekMissionSuccessRate;
+    private Double lastWeekMissionSuccessRate;
+    private Double thisWeekMissionSuccessRate;
 
 }
